% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{clean_data function}
\usage{
clean_data(recording, stim_sweep)
}
\arguments{
\item{recording, }{object generated from read_data}

\item{stim_sweep, }{sweep number at DHPG stim}
}
\value{
data object or list of objects
}
\description{
cleans data, takes two points and makes min average, subset data from 20 minutes before the stim as baseline,  calculates slope as % baseline returns clean object
}
