% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QC}
\alias{QC}
\title{QC function}
\usage{
QC(file_path, rig, stim_sweep, res_file)
}
\arguments{
\item{file_path, }{file path of XLS file}

\item{rig, }{Rig number}

\item{stim_sweep, }{sweep number at DHPG sti}

\item{res_file, }{output file name}
}
\value{
plot
}
\description{
plot
}
