% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCplot.R
\name{QCplot}
\alias{QCplot}
\title{QCplot function}
\usage{
QCplot(plt_data, res_file)
}
\arguments{
\item{plt_data, }{data to plot}

\item{res_file, }{output file name}
}
\value{
generates a plot
}
\description{
plot
}
