% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read_data function}
\usage{
read_data(file_path, rig)
}
\arguments{
\item{file_path, }{file path of XLS file}

\item{rig, }{Rig number}
}
\value{
data object or list of objects
}
\description{
This reads the XLS file generated from WinLTD software
}
